ajuste
cliente mega lojao

+++ integracao_odoo/pos_order_script.py
@@ -255,6 +255,7 @@ class IntegracaoOdoo:
                         'name': str(cod_forma),
                         'caixa': cod_caixa,
                         'cod_venda': cod_venda,
+                        'user_id': self.caixa_user,
             }
             arquivo_nome = self.path_envio + '/sangria_%s.json' %(cod_forma)
             if os.path.exists(arquivo_nome):
@@ -309,6 +310,7 @@ class IntegracaoOdoo:
             _logger.info("Sem produtos pra atualizar")
             return False
         for pr in prod_novo:
+            codproduto = pr['codproduto'] * 10
             ncm = ''
             if 'ncm' in pr and pr['ncm']:
                 ncm = pr['ncm']
@@ -335,33 +337,38 @@ class IntegracaoOdoo:
             codp = str(pr['codproduto'])
 
             sqlp = "select FIRST 1 CODPRODUTO,\
-                    DATACADASTRO, VALOR_PRAZO from produtos where codpro = '%s'" %(pr["codpro"])
+                    DATACADASTRO, VALOR_PRAZO from produtos \
+                    where codpro = '%s' \
+                    or codproduto = %s" %(pr["codpro"], str(codproduto))
             prods = db.query(sqlp)
             if pr['codpro']:
                 codpro = pr['codpro'].strip()
-            if not os.path.exists('promocao'):
-                os.makedirs('promocao')
-            promo_jpg = f"promocao\{codpro}.jpg"
-            promo_txt = f"promocao\{codpro}.txt"
-            img_sql = 'Null'
-            if pr['promocao_jpg']:
-                img_sql = promo_jpg
-                if not os.path.exists(promo_jpg):
-                    imagem = pr['promocao_jpg']
-                    imagem_dec = base64.b64decode(imagem)
-                    with open(promo_jpg, 'wb') as f:
-                        f.write(imagem_dec)
-            else:
-                if os.path.exists(promo_jpg):
-                    os.remove(promo_jpg)
-            if pr['promocao_txt']:
-                img_sql = promo_txt
-                if not os.path.exists(promo_txt):
-                    with open(promo_txt, 'w') as f:
-                        f.write(pr['promocao_txt'])
-            else:
-                if os.path.exists(promo_txt):
-                    os.remove(promo_txt)
+
+            # serve pra identificar se usa o pdv_integracao ou o pdv_integracao_outros
+            if 'tipo_venda' in pr:
+                if not os.path.exists('promocao'):
+                    os.makedirs('promocao')
+                promo_jpg = f"promocao{codpro}.jpg"
+                promo_txt = f"promocao{codpro}.txt"
+                img_sql = 'Null'
+                if pr['promocao_jpg']:
+                    img_sql = promo_jpg
+                    if not os.path.exists(promo_jpg):
+                        imagem = pr['promocao_jpg']
+                        imagem_dec = base64.b64decode(imagem)
+                        with open(promo_jpg, 'wb') as f:
+                            f.write(imagem_dec)
+                else:
+                    if os.path.exists(promo_jpg):
+                        os.remove(promo_jpg)
+                if pr['promocao_txt']:
+                    img_sql = promo_txt
+                    if not os.path.exists(promo_txt):
+                        with open(promo_txt, 'w') as f:
+                            f.write(pr['promocao_txt'])
+                else:
+                    if os.path.exists(promo_txt):
+                        os.remove(promo_txt)
 
             #sqlp = 'select codproduto from produtos where codpro like \'%s\'' %(codp+'%')
             if codbarra:
@@ -412,10 +419,12 @@ class IntegracaoOdoo:
                 #    codp = str(product_id.id) 
                 #print ('Incluindo - %s-%s' %(str(product_id.id),product_id.name))
                 un = pr['unidademedida'][:2]
-                codproduto = pr['codproduto'] * 10
+                
                 insere = 'INSERT INTO PRODUTOS (UNIDADEMEDIDA, PRODUTO, PRECOMEDIO, CODPRO,\
                           TIPOPRECOVENDA, ORIGEM, NCM, VALORUNITARIOATUAL, VALOR_PRAZO, TIPO, RATEIO, \
-                          QTDEATACADO, PRECOATACADO, DATACADASTRO, CODPRODUTO, FOTOPRODUTO'
+                          QTDEATACADO, PRECOATACADO, DATACADASTRO, CODPRODUTO'
+                if 'tipo_venda' in pr:
+                    insere += ', FOTOPRODUTO'
                 if codbarra:
                     insere += ', COD_BARRA'
 
@@ -435,7 +444,8 @@ class IntegracaoOdoo:
                 insere += ',' + precoatacado
                 insere += ', \'' + data_fb + '\''
                 insere += ',' + str(codproduto)
-                insere += ',' + img_sql                
+                if 'tipo_venda' in pr:
+                    insere += ',' + img_sql                
                 if codbarra:
                     insere += ', \'' + str(codbarra) + '\''
                 insere += ')'
@@ -473,10 +483,12 @@ class IntegracaoOdoo:
                 altera += ', QTDEATACADO = ' + qtdeatacado 
                 altera += ', PRECOATACADO = ' + precoatacado
                 altera += ', DATACADASTRO = \'' + data_fb + '\''
-                altera += ', FOTOPRODUTO = \'' + img_sql + '\''
+                if 'tipo_venda' in pr:
+                    altera += ', FOTOPRODUTO = \'' + img_sql + '\''
                 if codbarra:
                     altera += ', COD_BARRA = \'' + str(codbarra) + '\''
                 altera += ' WHERE CODPRO = \'' + str(pr['codpro']) + '\''
+                altera += ' OR CODPRODUTO = ' + str(codproduto)
                 retorno = db.insert(altera)
                 if retorno:
                     print ('SQL : %s' %(altera))
@@ -503,11 +515,12 @@ class IntegracaoOdoo:
 
     def action_atualiza_clientes(self):
         self._inicia()
-        try:
-            db = con()
-            dblocal = local_db(filename = 'lancamento.db', table = 'cliente')
-        except:
-            msg_sis = u'Caminho ou nome do banco invalido. '
+        envia("cliente")
+        # try:
+        db = con()
+        #     dblocal = local_db(filename = 'lancamento.db', table = 'cliente')
+        # except:
+        #     msg_sis = u'Caminho ou nome do banco invalido. '
         msg_erro = ''
         msg_sis = 'Integrando Clientes para o PDV '
         print("Atualizando clientes.")
@@ -515,11 +528,20 @@ class IntegracaoOdoo:
         alterado = 0
         hj = datetime.now()
         hj = datetime.strftime(hj,'%Y-%m-%d')
-        cli_ids = dblocal.consulta_cliente()
-        if not cli_ids:
+        # cli_ids = dblocal.consulta_cliente()
+        arquivo_json = os.path.join(self.path_envio, "clientes.json")
+        try:
+            f = open(arquivo_json)
+        except:
+            return
+        cliente_arquivo = json.load(f)
+        
+        cli_novo = json.loads(cliente_arquivo['result'])
+
+        if not cli_novo:
             print('Sem clientes pra atualizar')
             return
-        for partner_id in cli_ids:
+        for partner_id in cli_novo:
             nome = partner_id['nomecliente']
             nome = self.removerAcentosECaracteresEspeciais(nome)
             razao = nome
